Allow only one database provider per service provider
Issue #7457

Note that the way AddEntityFramework works has changed. Previously, the provider indirection mechanism ensured that any service overridden by a provider would be honored. Since this is gone, we use the following approach:
- TryAdd all provider services
- TryAdd all relational services (any services already registered by the provider are not overwritten)
- TryAdd all core services (any services registered by first two steps are not overwritten)
- Any services explicitly added by application will remain regardless

This means that an application should never call AddEntityFramework directly, since if this is done before the provider services are added then things will break. Therefore this method is renamed and no-longer an extension method. (Note that applications should not have been calling it anyway, but doing so was previously harmless.)

Also note that all services registered are now interfaces. This should have been the case before except for provider-specific implementations that were still resolved through their interfaces, but some concrete services had crept in.

Work still to do:
- Revisit the scopes of services
- Consider combining core convention set builder and model-specific builder into one
- Find a new solution for InjectAdditionalServices, since there is now no longer a hook for this
- Look at ResultOperatorHandler, since it differs from all other services in how it is used